fix: hedera_format_x is off by one, decrement reverse index after initial operation
