+++ src/hedera.c
@@ -1,9 +1,7 @@
 #include <os.h>
 #include <cx.h>
 #include "globals.h"
-#include "printf.h"
 #include "hedera.h"
-#include "string.h"
 
 bool hedera_derive_keypair(
     uint32_t index,
@@ -103,47 +101,3 @@ bool hedera_sign(
 
     return true;
 }
-
-char* hedera_format_tinybar(uint64_t tinybar) {
-    return hedera_format_amount(tinybar, 8);
-}
-
-char* hedera_format_amount(uint64_t amount, uint8_t decimals) {
-    static const int BUF_SIZE = 32;
-    static char buf[BUF_SIZE];
-
-    explicit_bzero(buf, BUF_SIZE);
-
-    int i = 0;
-
-    while (i < (BUF_SIZE - 1) && (amount > 0 || i < decimals)) {
-        int digit = amount % 10;
-        amount /= 10;
-
-        buf[i++] = '0' + digit;
-
-        if (i == decimals) {
-            buf[i++] = '.';
-        }
-    }
-
-    if (buf[i - 1] == '.') {
-        buf[i++] = '0';
-    }
-
-    int j = 0;
-    char tmp;
-
-    while (j < i) {
-        j += 1;
-        i -= 1;
-
-        tmp = buf[j];
-        buf[j] = buf[i];
-        buf[i] = tmp;
-    }
-
-    return buf;
-}
-
-
